- Win: fixed compilation after introduction of the new parser services. - Win: table editor wrapper: removed obsolete get_all_triggers_sql(). - Some cleanup on the way.

+++ modules/db.mysql.parser/src/mysql_parser_module.cpp
@@ -66,7 +66,7 @@ int MySQLParserServicesImpl::checkSqlSyntax(ParserContext::Ref context, const ch
                                             size_t length, MySQLQueryType type)
 {
   context->recognizer()->parse(sql, length, true, type);
-  return context->recognizer()->error_info().size();
+  return (int)context->recognizer()->error_info().size();
 }
 
 //--------------------------------------------------------------------------------------------------
